pyproject.toml
[project]
name = "dive_color_corrector"
dynamic = ["version"]
description = "A tool for correcting underwater images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "tensorflow-cpu>=2.15.0",
    "keras>=2.15.0",
]

[project.optional-dependencies]
gui = [
    "PySimpleGUI>=4.60.0",
]

[build-system]
requires = ["hatchling>=1.18.0", "uv>=0.6.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dive_color_corrector/__init__.py"

[tool.hatch.build]
packages = ["src/dive_color_corrector"]

[tool.hatch.build.targets.wheel]
packages = ["src/dive_color_corrector"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "RUF"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["dive_color_corrector"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
